refactor: cursos - add validação nas server actions

+++ src/actions/cursos/delete-cursos.ts
@@ -1,12 +1,28 @@
 "use server";
 
+import { getServerSession } from "next-auth";
+
+import { authOptions } from "@/lib/auth";
 import prisma from "@/lib/prisma";
 
 export const deleteCurso = async (idCurso: number) => {
-  const curso = await prisma.curso.delete({
-    where: {
-      idCurso: Number(idCurso),
-    },
-  });
-  return curso;
+  try {
+    const session = await getServerSession(authOptions);
+
+    if (!session || session.user.role !== "admin")
+      return { error: "Usuário não autorizado" };
+
+    if (!idCurso)
+      return { error: "Erro ao excluir dados; item não selecionado." };
+
+    const curso = await prisma.curso.delete({
+      where: {
+        idCurso: Number(idCurso),
+      },
+    });
+    return curso;
+  } catch (error) {
+    console.error("Erro ao excluir curso: ", error);
+    return { error: "Erro interno do servidor" };
+  }
 };
